+++ example_test.go
@@ -10,8 +10,8 @@ import (
 
 func ExampleConfigure() {
 	s := sequencer.NewMillisecond() // has 4 bytes free space for a node
-	n := uint(19)                   // Base62 => J
-	t := uint(0)                    // initial time (start from unix time in ms)
+	n := uint64(19)                 // Base62 => J
+	t := uint64(0)                  // initial time (start from unix time in ms)
 
 	monoton.Configure(s, n, t)
 	fmt.Println(monoton.Next()[12:])
@@ -20,9 +20,9 @@ func ExampleConfigure() {
 }
 
 func ExampleNext() {
-	s := sequencer.NewSecond()   // sequencer.Second
-	n := uint(19)                // Base62 => J
-	t := uint(time.Now().Unix()) // initial time (start from unix time in s)
+	s := sequencer.NewSecond()     // sequencer.Second
+	n := uint64(19)                // Base62 => J
+	t := uint64(time.Now().Unix()) // initial time (start from unix time in s)
 
 	monoton.Configure(s, n, t)
 	fmt.Println(len(monoton.Next()))
