+++ sequencer/sequence_test.go
@@ -8,7 +8,7 @@ import (
 )
 
 func TestMax_Sequence(t *testing.T) {
-	want := uint(12345)
+	want := uint64(12345)
 	s := &Sequence{max: want}
 
 	if got := s.Max(); got != want {
@@ -17,7 +17,7 @@ func TestMax_Sequence(t *testing.T) {
 }
 
 func TestMaxTime_Sequence(t *testing.T) {
-	want := uint(1<<64 - 1)
+	want := uint64(1<<64 - 1)
 	s := &Sequence{maxTime: want}
 
 	if got := s.MaxTime(); got != want {
@@ -27,11 +27,11 @@ func TestMaxTime_Sequence(t *testing.T) {
 
 func TestNext_Sequence(t *testing.T) {
 	tests := []struct {
-		want uint
-		now  func() uint
+		want uint64
+		now  func() uint64
 	}{
-		{want: uint(2), now: func() uint { return 0 }},
-		{want: uint(0), now: func() uint { time.Sleep(time.Nanosecond); return mtimer.Now() }},
+		{want: uint64(2), now: func() uint64 { return 0 }},
+		{want: uint64(0), now: func() uint64 { time.Sleep(time.Nanosecond); return mtimer.Now() }},
 	}
 
 	for _, test := range tests {
